refactor: move window files to src/gui/windows/ subdirectory
