+++ src/gui/windows/dde_status_renderer.h
@@ -1,0 +1,17 @@
+#pragma once
+
+#include <functional>
+
+#include "gui/constants.h"
+#include "logger/logger.h"
+#include "dde/client.h"
+
+class ZemaxDDEClient;
+
+namespace gui {
+    class DdeStatusRenderer {
+    public:
+        void renderDdeStatus(ZemaxDDE::ZemaxDDEClient* ddeClient, Logger& logger);
+        void renderDdeStatusFrame(ZemaxDDE::ZemaxDDEClient* ddeClient, Logger& logger);
+    };
+}
