+++ src/gui/windows/debug_log_viewer.h
