+++ src/gui/windows/optical_system_info.cpp
